Implement initial loading support for iml-gui
Fixes #1528.

  - Remove home route.
  - extract a few pieces from lib.
  - implement loading component.
