--- iml-gui/crate/src/page/home.rs
@@ -1,6 +1,0 @@
-use crate::{Model, Msg};
-use seed::{prelude::*, *};
-
-pub fn view(_model: &Model) -> impl View<Msg> {
-    div![]
-}
