+++ iml-gui/entries/templates/loading_page.hbs
@@ -1,7 +1,6 @@
-<div id="loading-page" class="fade-in h-screen flex flex-col items-center justify-center">
-    <h1 class="-mt-10vh font-sans font-light text-29 text-gray-10 sm:text-45 lg:text-80">Loading...</h1>
+<div id="loading-page" class="fade-in h-screen flex flex-col items-center justify-center text-gray-500">
     {{! spinner from https://loading.io }}
-    <svg class="mt-4 h-12 rotate sm:mt-5 sm:h-16 lg:mt-8 lg:h-24" xmlns="http://www.w3.org/2000/svg"
+    <svg class="mt-4 h-12 rotate sm:mt-5 sm:h-16 lg:mt-8 lg:h-24 fill-current" xmlns="http://www.w3.org/2000/svg"
         viewBox="0 0 100 100" preserveAspectRatio="xMidYMid"
         style="background: rgba(0, 0, 0, 0) none repeat scroll 0% 0%;">
         <path stroke="none" d="M10 50A40 40 0 0 0 90 50A40 42 0 0 1 10 50">
